Moved spec attachment data to fixture
